+++ AsanaNet/Models/AsanaResponse.cs
@@ -6,4 +6,7 @@ public class AsanaResponse<T>
 {
     [JsonPropertyName("data")]
     public T? Data { get; set; }
+
+    [JsonPropertyName("errors")]
+    public List<AsanaError>? Errors { get; set; }
 }